+++ Changelog.md
@@ -28,6 +28,7 @@ Bugfixes:
  * SMTChecker: Fix false positive in external calls from constructors.
  * SMTChecker: Fix internal error on some multi-source uses of ``abi.*``, cryptographic functions and constants.
  * SMTChecker: Fix BMC's constraints regarding internal functions.
+ * Standard JSON: Fix non-fatal errors in Yul mode being discarded if followed by a fatal error.
  * Type Checker: Disallow modifier declarations and definitions in interfaces.
  * Yul Optimizer: Fix a crash in LoadResolver, when ``keccak256`` has particular non-identifier arguments.